pyproject.toml
[project]
name = "surya-sidhanta"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pyswisseph>=2.10.3.2",
    "scipy>=1.16.3",
    "tabulate>=0.9.0",
    "tqdm>=4.67.1",
]
